backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
discord.py==2.3.2
dj-rest-auth==6.0.0
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-steam-api==1.2.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
